// ui/src/generated/OperationSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CapabilityRef } from "./CapabilityRef";
import type { OperationId } from "./OperationId";
import type { OperationStatus } from "./OperationStatus";

export type OperationSummary = { cursor: number, operation_id: OperationId, client_request_id: string, capability: CapabilityRef, status: OperationStatus, accepted_at_ms: number, updated_at_ms: number, error: string | null, };